pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdwpf"
version = "0.1.0"
description = "SDWPF loaders, features, and wind power forecasting helpers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
experiments = ["mlflow>=2.10"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
